Conflicts:

	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/expressorder/ExpressOrderDataService_driver.class

	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/expressorder/ExpressOrderDataService_stub.class

+++ ExpressManageSystemGAP/common/src/main/java/gap/common/po/RentPO.java
@@ -4,21 +4,21 @@ import java.util.Date;
 
 public class RentPO {
 
-	private int money;
+	private double money;
 	private Date lastPaidDate;
     private String institution;
    
-	public RentPO(int money, Date lastPaidDate,String institution) {
+	public RentPO(double money, Date lastPaidDate,String institution) {
 		this.setMoney(money);
 		this.setLastPaidDate(lastPaidDate);
 		this.setInstitution(institution);
 	}
 
-	public int getMoney() {
+	public double getMoney() {
 		return money;
 	}
 
-	public void setMoney(int money) {
+	public void setMoney(double money) {
 		this.money = money;
 	}
